+++ libs/poses/src/CPosePDFGaussianInf.cpp
@@ -264,7 +264,7 @@ void CPosePDFGaussianInf::drawManySamples(
 
 	getRandomGenerator().drawGaussianMultivariateMany(rndSamples, N, cov);
 	outSamples.resize(N);
-	for (unsigned int i = 0; i < N; i++)
+	for (size_t i = 0; i < N; i++)
 	{
 		outSamples[i].resize(3);
 		outSamples[i][0] = mean.x() + rndSamples[i][0];
